Cleanup: separate Excel formats into modules
